[ui] Split AssetNodeOverview into sub-components
The AssetNodeOverview component was approaching 1000 LOC and assembles a lot of metadata, etc. at the top that is far from where it’s rendered.  This PR doesn’t contain any logic changes, I just broke the file up so that the metadata being pulled out of the props is closer to where it’s used and the whole thing is a bit easier to grok.
